fix: prevent re-backup symlink on Windows

+++ scripts/nvim/enable.ps1
@@ -2,7 +2,7 @@ $NVIM_SOURCE = "$pwd\nvim"
 $NVIM_DESTINATION = "$($env:LOCALAPPDATA)\nvim"
 $NVIM_BACKUP = "$($env:LOCALAPPDATA)\nvim_backup"
 
-if (Test-Path -Path $NVIM_DESTINATION){
+if ( (Test-Path -Path $NVIM_DESTINATION) -and ((Get-Item $NVIM_DESTINATION).LinkType -eq $null) ){
 	echo "Creating backup directory for nvim..."
 	Move-Item -Path $NVIM_DESTINATION -Destination $NVIM_BACKUP
 }
